requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.9.1
blessed==1.21.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2
django-cors-headers==4.7.0
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
jinxed==1.3.0
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==3.5.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20250822
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
